move scienceBeamApi.js to xpub-controller
