+++ server/xpub-controller/client/scienceBeamApi.js
