+++ server/xpub-server/entities/manuscript/resolvers/uploadManuscript.test.js
@@ -12,7 +12,7 @@ const { createTables } = require('@pubsweet/db-manager')
 const mailer = require('@pubsweet/component-send-email')
 const startS3rver = require('../../../test/mock-s3-server')
 const { User, Manuscript } = require('@elifesciences/xpub-model')
-const scienceBeamApi = require('./scienceBeamApi')
+const scienceBeamApi = require('../../../../xpub-controller/client/scienceBeamApi')
 const { Mutation } = require('.')
 const { userData, badUserData } = require('./index.test.data')
 const replaySetup = require('../../../../../test/helpers/replay-setup')
